requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.12.0
databases[postgresql]>=0.8.0
asyncpg>=0.29.0
pydantic-settings>=2.0.0
httpx>=0.25.0
cryptography>=40.0.0
python-dotenv>=1.0.0
